# docs/swagger.yml
swagger: '2.0'
info:
  title: Quicksilver Chain - REST API
  description: REST interface for query and transaction services
  version: 1.0.0
paths:
...
  /quicksilver/tx/v1/interchainstaking/close_channel:
    post:
      summary: >-
        SignalIntent defines a method for signalling voting intent for one or
        more

        validators.
      operationId: GovCloseChannel
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            description: >-
              MsgGovCloseChannelResponse defines the MsgGovCloseChannel response
              type.
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    type_url:
                      type: string
                    value:
                      type: string
                      format: byte
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              title:
                type: string
              description:
                type: string
              channel_id:
                type: string
              port_id:
                type: string
              authority:
                type: string
      tags:
        - Msg
  /quicksilver/tx/v1/interchainstaking/intent:
    post:
      summary: >-
        SignalIntent defines a method for signalling voting intent for one or
        more

        validators.
      operationId: SignalIntent
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            description: MsgSignalIntentResponse defines the MsgSignalIntent response type.
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    type_url:
                      type: string
                    value:
                      type: string
                      format: byte
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              chain_id:
                type: string
              intents:
                type: string
              from_address:
                type: string
            description: >-
              MsgSignalIntent represents a message type for signalling voting
              intent for

              one or more validators.
      tags:
        - Msg
  /quicksilver/tx/v1/interchainstaking/redeem:
    post:
      summary: |-
        RequestRedemption defines a method for requesting burning of qAssets for
        native assets.
      operationId: RequestRedemption
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            description: >-
              MsgRequestRedemptionResponse defines the MsgRequestRedemption
              response type.
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    type_url:
                      type: string
                    value:
                      type: string
                      format: byte
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              value:
                type: object
                properties:
                  denom:
                    type: string
                  amount:
                    type: string
                description: >-
                  Coin defines a token with a denomination and an amount.


                  NOTE: The amount field is an Int which implements the custom
                  method

                  signatures required by gogoproto.
              destination_address:
                type: string
              from_address:
                type: string
            description: >-
              MsgRequestRedemption represents a message type to request a burn
              of qAssets

              for native assets.
      tags:
        - Msg
  /quicksilver/tx/v1/interchainstaking/reopen_channel:
    post:
      operationId: GovReopenChannel
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            description: >-
              MsgGovReopenChannelResponse defines the MsgGovReopenChannel
              response type.
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    type_url:
                      type: string
                    value:
                      type: string
                      format: byte
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              title:
                type: string
              description:
                type: string
              connection_id:
                type: string
              port_id:
                type: string
              authority:
                type: string
      tags:
        - Msg
...